chatgpt/chatgpt/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow>=10.0.0
